risk level logic added

+++ Models/TestResults.cs
@@ -33,7 +33,7 @@ namespace EntraGraphAPI.Models
         public int? denyThreshold { get; set; }
         public int? permitCount { get; set; }
         public int? accessCount { get; set; }
-        public float? final_trust_factor { get; set; }
+        public decimal? final_trust_factor { get; set; }
         public bool? final_result { get; set; }
         public int? test_run_id { get; set; }
         public string? risk_level { get; set; }
